backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
h11==0.14.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
groq==0.11.0
sqlparse==0.5.1
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
PyMySQL==1.1.1
cryptography==43.0.1
firebase-admin==6.5.0
httpx==0.27.2
